fix: update keys in components
This commit fixes an issue in the `CaptionSelectionPopout` and `CaptionSettingsPopout` components where the `key` prop for the `PopoutListEntry` and `CaptionColorSelector` components respectively was not unique. The `key` prop has been updated from using `link.langIso` and `color` to `link.id` and `color` respectively to ensure uniqueness and prevent potential rendering issues.

Additionally, in the `SettingsModal` component, the `key` prop for `CaptionColorSelector` has also been updated to use `color` for consistency.